requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.6
fastapi-cli==0.0.5
fastjsonschema==2.19.1
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
kanren==0.2.3
keras==3.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
llama-cloud==0.1.9
llama-index==0.12.11
llama-index-agent-openai==0.4.2
llama-index-cli==0.4.0
llama-index-core==0.12.11
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.13
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.3
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
openai==1.59.7
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.14
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==26.0.3
qrcode==8.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.18.1
ruff==0.9.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.0
tomlkit==0.13.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.67
ultralytics-thop==2.0.14
unification==0.2.2
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.28.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.43.0
whitenoise==6.8.2
wrapt==1.16.0
yarl==1.18.3
